ubusd_main: check strdup return value in mkdir_sockdir
strdup() can fail and return NULL; the subsequent strrchr() call then
dereferences a NULL pointer.  The impact is limited to startup with a
constant path, but handle the allocation failure gracefully.

Link: https://github.com/openwrt/ubus/pull/20

+++ ubusd_main.c
@@ -271,17 +271,32 @@ static struct uloop_fd sighup_ufd = {
 	.cb = sighup_fd_cb,
 };
 
-static void mkdir_sockdir()
+static int mkdir_sockdir()
 {
 	char *ubus_sock_dir, *tmp;
+	int ret = 0;
 
 	ubus_sock_dir = strdup(UBUS_UNIX_SOCKET);
+	if (!ubus_sock_dir)
+		return -1;
+
 	tmp = strrchr(ubus_sock_dir, '/');
 	if (tmp) {
 		*tmp = '\0';
-		mkdir(ubus_sock_dir, 0755);
+		ret = mkdir(ubus_sock_dir, 0755);
+		if (ret && errno == EEXIST) {
+			struct stat st;
+
+			if (stat(ubus_sock_dir, &st) == 0 && S_ISDIR(st.st_mode))
+				ret = 0;
+			else
+				fprintf(stderr, "%s exists but is not a directory\n",
+					ubus_sock_dir);
+		}
 	}
+
 	free(ubus_sock_dir);
+	return ret;
 }
 
 #include <libubox/ulog.h>
@@ -332,7 +347,9 @@ int main(int argc, char **argv)
 		}
 	}
 
-	mkdir_sockdir();
+	ret = mkdir_sockdir();
+	if (ret)
+		goto out;
 	unlink(ubus_socket);
 	umask(0111);
 	server_fd.fd = usock(USOCK_UNIX | USOCK_SERVER | USOCK_NONBLOCK, ubus_socket, NULL);
